backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
ariadne==0.23.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.14.0
eval_type_backport==0.2.0
faiss-cpu==1.9.0
fastapi==0.115.3
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-vision==3.8.0
googleapis-common-protos==1.65.0
graphql-core==3.2.5
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
inflection==0.5.1
iopath==0.1.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.137
layoutparser==0.3.4
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime
openai==1.52.2
openapi==2.0.0
opencv-python==4.10.0.84
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.20.0
pikepdf==9.4.0
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scipy
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
gunicorn==21.2.0
SQLAlchemy==2.0.36
starlette==0.41.0
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.4
unstructured-client==0.26.2
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.32.0
webencodings==0.5.1
wrapt==1.16.0
yarl==1.16.0
youtube-transcript-api==0.6.3
yt-dlp==2025.3.21